refactor(test): update imports and GameState to G model
- Update imports from src.balatrobot to balatrobot package structure
- Replace GameState references with G model throughout test files
- Update test class documentation and comments to reflect new naming
